pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workspace-mcp"
version = "1.19.0"
description = "Gmail-only MCP server (stripped from google_workspace_mcp)."
readme = "README.md"
keywords = [ "mcp", "google", "gmail", "llm", "ai", "claude", "model", "context", "protocol", "server"]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
 "fastapi>=0.115.12",
 "fastmcp>=3.2.0",
 "google-api-python-client>=2.168.0",
 "google-auth-oauthlib>=1.2.2",
 "httpx>=0.28.1",
 "py-key-value-aio>=0.3.0",
 "pyjwt>=2.12.0",
 "python-dotenv>=1.1.0",
 "pyyaml>=6.0.2",
 "cryptography>=45.0.0",
 "defusedxml>=0.7.1",
 "pypdf>=6.10.2",
]

[project.scripts]
workspace-mcp = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.3",
    "ruff>=0.12.4",
]

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "build", "dist"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.setuptools.package-data]
core = ["tool_tiers.yaml"]
